+++ .npmignore
@@ -1,0 +1,2 @@
+dist/integration-test
+dist/unit-test
