+++ ts-src/integration-test/db.ts
@@ -1,4 +1,4 @@
-import { pool, Connection } from '../lib/jt400'
+import { pool, Connection } from '..'
 export const jt400: Connection = pool({
   'date format': 'iso',
 })
