+++ ts-src/lib/baseConnection.types.ts
