+++ ts-src/lib/connection.types.ts
