+++ ts-src/lib/ifs/read_stream.ts
