+++ ts-src/lib/ifs/write_stream.ts
