+++ ts-src/lib/jdbcstream.ts
