+++ ts-src/lib/jdbcwritestream.ts
