+++ ts-src/lib/sqlutil.ts
